pyproject.toml
[project]
name = "fastapi-cli"
dynamic = ["version"]
description = "Run and manage FastAPI apps from the command line with FastAPI CLI. 🚀"
authors = [
    {name = "Sebastián Ramírez", email = "tiangolo@gmail.com"},
]
requires-python = ">=3.8"
readme = "README.md"
license = {text = "MIT"}
classifiers = [
    "Intended Audience :: Information Technology",
    "Intended Audience :: System Administrators",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python",
    "Topic :: Software Development :: Libraries :: Application Frameworks",
    "Topic :: Software Development :: Libraries :: Python Modules",
    "Topic :: Software Development :: Libraries",
    "Topic :: Software Development",
    "Typing :: Typed",
    "Development Status :: 4 - Beta",
    "Framework :: FastAPI",
    "Intended Audience :: Developers",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3 :: Only",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "License :: OSI Approved :: MIT License",
]
dependencies = [
    "typer >= 0.15.1",
    "uvicorn[standard] >= 0.15.0",
    "rich-toolkit >= 0.14.8",
    "pydantic-settings",
]

[project.optional-dependencies]
standard = [
    "uvicorn[standard] >= 0.15.0",
    "fastapi-cloud-cli >= 0.1.1",
]
standard-no-fastapi-cloud-cli = [
    "uvicorn[standard] >= 0.15.0",
]

[project.urls]
Homepage = "https://github.com/fastapi/fastapi-cli"
Documentation = "https://fastapi.tiangolo.com/fastapi-cli/"
Repository = "https://github.com/fastapi/fastapi-cli"
Issues = "https://github.com/fastapi/fastapi-cli/issues"
Changelog = "https://github.com/fastapi/fastapi-cli/blob/main/release-notes.md"

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
version = { source = "file", path = "src/fastapi_cli/__init__.py" }
distribution = true

[tool.pdm.build]
source-includes = [
    "tests/",
    "requirements*.txt",
    "scripts/",
    ]

[tool.tiangolo._internal-slim-build.packages.fastapi-cli-slim.project]
name = "fastapi-cli-slim"

[tool.pytest.ini_options]
addopts = [
  "--strict-config",
  "--strict-markers",
]
xfail_strict = true
junit_family = "xunit2"

[tool.coverage.run]
parallel = true
data_file = "coverage/.coverage"
source = [
    "src",
    "tests",
]
context = '${CONTEXT}'
dynamic_context = "test_function"
omit = [
    "tests/assets/*",
]

[tool.coverage.report]
show_missing = true
sort = "-Cover"
exclude_lines = [
    "pragma: no cover",
    "@overload",
    'if __name__ == "__main__":',
    "if TYPE_CHECKING:",
]

[tool.coverage.html]
show_contexts = true

[tool.mypy]
strict = true
exclude = [
    "tests/assets/*",
]

[tool.ruff.lint]
select = [
    "E",  # pycodestyle errors
    "W",  # pycodestyle warnings
    "F",  # pyflakes
    "I",  # isort
    "B",  # flake8-bugbear
    "C4",  # flake8-comprehensions
    "UP", # pyupgrade
]
ignore = [
    "E501",  # line too long, handled by black
    "B008",  # do not perform function calls in argument defaults
    "C901",  # too complex
    "W191", # indentation contains tabs
]

# [tool.ruff.lint.per-file-ignores]
# "__init__.py" = ["F401"]


[tool.ruff.lint.isort]
known-third-party = ["typer", "fastapi"]

[tool.ruff.lint.pyupgrade]
# Preserve types, even if a file imports `from __future__ import annotations`.
keep-runtime-typing = true
